backend/pyproject.toml
[tool.poetry]
name = "resume-parser"
version = "0.1.0"
description = ""
authors = ["sortofasian <charliesyvertsen06@icloud.com>"]

[tool.poetry.dependencies]
python = "3.8.13"
gensim = "^4.2.0"
Flask = "^2.1.2"
textract = "^1.6.5"
requests = "^2.28.1"
pymongo = "^4.1.1"
nanoid = "^2.0.0"
bcrypt = "^3.2.2"
dnspython = "^2.2.1"
PyJWT = "^2.4.0"
flake8 = "^4.0.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
jupyterlab = "^3.4.3"
flake8 = "^4.0.1"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
